refactor(assets): extract shared SearchDropdown behavior from search/top-search
search_controller.js and top_search_controller.js had accumulated
near-identical debounce/fetch/keyboard-nav/dropdown-visibility code
independently (19 matching method names). Factored the shared parts
into a SearchDropdown mixin (assets/js/search-dropdown.js); each
controller now only defines its own URL building, result rendering,
and selection behavior.

Two small behavior deltas fall out of using one implementation
instead of two, both strict improvements: top-search now gets
search_controller's accent-insensitive highlight matching (was a
plain substring match), and a defensive `data.error` check that
search_controller already had.
