+++ assets/controllers/search_controller.js
@@ -1,28 +1,24 @@
 import { Controller } from '@hotwired/stimulus';
 import { trans } from '../translator';
+import { SearchDropdown } from '../js/search-dropdown';
+
 /**
- * Modern Search Controller - Replaces legacy typeahead.js implementation
- *
- * Responsibilities:
- * - Handle user input with debouncing (300ms delay)
- * - Send AJAX requests to search API
- * - Manage search dropdown visibility and state
- * - Handle keyboard navigation (arrow keys, enter, escape)
- * - Manage "Show more results" functionality
- * - Provide accessibility features with ARIA labels
+ * Site-wide search bar -- shared debounce/fetch/keyboard-nav/dropdown
+ * behavior lives in SearchDropdown (see assets/js/search-dropdown.js);
+ * this defines what's specific to searching coasters/parks/users and
+ * navigating to the selected result.
  *
  * Usage:
  * <div data-controller="search"
  *      data-search-search-url-value="/api/search"
  *      data-search-results-url-value="/search">
- *   <input data-search-target="input" data-action="input->search#search keydown->search#handleKeydown">
+ *   <input data-search-target="input" data-action="input->search#handleInput keydown->search#handleKeydown">
  *   <div data-search-target="dropdown">
  *     <div data-search-target="results"></div>
  *   </div>
  * </div>
  */
-export default class extends Controller {
-    static targets = ['input', 'dropdown', 'results'];
+export default class extends SearchDropdown(Controller) {
     static values = {
         searchUrl: String,
         resultsUrl: String,
@@ -31,165 +27,24 @@ export default class extends Controller {
         maxResults: { type: Number, default: 5 },
     };
 
+    get dropdownAriaLabel() {
+        return 'Search suggestions';
+    }
+
+    get loadingText() {
+        return trans('app.search.searching');
+    }
+
     connect() {
-        // Initialize state
-        this.debounceTimer = null;
-        this.currentRequest = null;
-        this.selectedIndex = -1;
-        this.isOpen = false;
-        this.lastQuery = '';
-
-        // Set up accessibility attributes
-        this.setupAccessibility();
-
-        // Set up event listeners
-        this.setupEventListeners();
-
-        // Initialize clear button visibility
+        super.connect();
         this.updateClearButtonVisibility();
     }
 
-    disconnect() {
-        // Clean up timers and requests
-        if (this.debounceTimer) {
-            clearTimeout(this.debounceTimer);
-        }
-
-        if (this.currentRequest) {
-            this.currentRequest.abort();
-        }
-
-        // Remove global event listeners
-        document.removeEventListener('click', this.closeOnOutsideClick);
-    }
-
-    /**
-     * Set up accessibility attributes for screen readers
-     */
-    setupAccessibility() {
-        if (this.hasInputTarget) {
-            this.inputTarget.setAttribute('role', 'combobox');
-            this.inputTarget.setAttribute('aria-expanded', 'false');
-            this.inputTarget.setAttribute('aria-autocomplete', 'list');
-            this.inputTarget.setAttribute('aria-haspopup', 'listbox');
-        }
-
-        if (this.hasDropdownTarget) {
-            this.dropdownTarget.setAttribute('role', 'listbox');
-            this.dropdownTarget.setAttribute(
-                'aria-label',
-                'Search suggestions'
-            );
-        }
-    }
-
-    /**
-     * Set up event listeners
-     */
-    setupEventListeners() {
-        // Close dropdown when clicking outside
-        this.closeOnOutsideClick = (event) => {
-            if (this.element && !this.element.contains(event.target)) {
-                this.hideDropdown();
-            }
-        };
-        document.addEventListener('click', this.closeOnOutsideClick);
-    }
-
-    /**
-     * Handle search input with debouncing
-     */
-    search(event) {
-        const query = event.target.value.trim();
-
-        // Clear previous timer
-        if (this.debounceTimer) {
-            clearTimeout(this.debounceTimer);
-        }
-
-        // Hide dropdown if query is too short
-        if (query.length < this.minLengthValue) {
-            this.hideDropdown();
-            return;
-        }
-
-        // Don't search if query hasn't changed
-        if (query === this.lastQuery) {
-            return;
-        }
-
-        // Debounce the search
-        this.debounceTimer = setTimeout(() => {
-            this.performSearch(query);
-        }, this.debounceDelayValue);
-    }
-
-    /**
-     * Handle input focus to clear placeholder behavior
-     */
-    handleFocus(event) {
-        // Ensure placeholder is properly handled
-        const input = event.target;
-        if (input.value.trim().length >= this.minLengthValue) {
-            this.search(event);
-        }
-    }
-
-    /**
-     * Execute AJAX search request
-     */
-    async performSearch(query) {
-        // Cancel previous request
-        if (this.currentRequest) {
-            this.currentRequest.abort();
-        }
-
-        this.lastQuery = query;
-
-        try {
-            // Show loading state
-            this.showLoadingState();
-
-            // Create AbortController for request cancellation
-            const controller = new AbortController();
-            this.currentRequest = controller;
-
-            const url = new URL(this.searchUrlValue, window.location.origin);
-            url.searchParams.set('q', query);
-            url.searchParams.set('limit', this.maxResultsValue.toString());
-
-            const response = await fetch(url.toString(), {
-                method: 'GET',
-                headers: {
-                    Accept: 'application/json',
-                    'X-Requested-With': 'XMLHttpRequest',
-                },
-                signal: controller.signal,
-            });
-
-            if (!response.ok) {
-                throw new Error(`HTTP error! status: ${response.status}`);
-            }
-
-            const data = await response.json();
-
-            if (data.error) {
-                throw new Error(data.message || 'Search error occurred');
-            }
-
-            this.updateResults(data);
-            this.showDropdown();
-        } catch (error) {
-            if (error.name === 'AbortError') {
-                // Request was cancelled, ignore
-                return;
-            }
-
-            console.error('Search error:', error);
-            this.showErrorState(error.message);
-        } finally {
-            this.currentRequest = null;
-        }
+    buildSearchUrl(query) {
+        const url = new URL(this.searchUrlValue, window.location.origin);
+        url.searchParams.set('q', query);
+        url.searchParams.set('limit', this.maxResultsValue.toString());
+        return url.toString();
     }
 
     /**
@@ -198,7 +53,7 @@ export default class extends Controller {
     updateResults(data) {
         if (!this.hasResultsTarget) return;
 
-        const { results, totalResults, hasMore, query } = data;
+        const { results, hasMore, query } = data;
         let html = '';
         let allResults = [];
 
@@ -228,21 +83,18 @@ export default class extends Controller {
             const hasMoreResults =
                 hasMore || allResults.length > maxDisplayResults;
 
-            // Render the limited results
             displayResults.forEach((item, index) => {
                 html += this.renderResultItem(item, index, query);
             });
 
-            // Add "Show more results" if there are more results (as 6th item)
             if (hasMoreResults) {
                 html += this.renderShowMoreOption(query);
             }
         }
 
         this.resultsTarget.innerHTML = html;
-        this.selectedIndex = -1; // Reset selection
+        this.selectedIndex = -1;
 
-        // Set up click handlers for results
         this.setupResultClickHandlers();
     }
 
@@ -296,47 +148,7 @@ export default class extends Controller {
             .replace(/^\w/, (c) => c.toUpperCase());
     }
 
-    /**
-     * Highlight search terms in text with accent-insensitive matching
-     */
-    highlightSearchTerm(text, query) {
-        if (!text || !query) return this.escapeHtml(text);
-
-        const escapedText = this.escapeHtml(text);
-
-        // Normalize text for accent-insensitive comparison
-        const normalizeText = (str) =>
-            str.normalize('NFD').replace(/[\u0300-\u036f]/g, '');
-
-        const normalizedText = normalizeText(text.toLowerCase());
-        const normalizedQuery = normalizeText(query.toLowerCase());
-
-        const matchIndex = normalizedText.indexOf(normalizedQuery);
-
-        if (matchIndex === -1) {
-            return escapedText;
-        }
-
-        // Get the actual matching text from original (with accents)
-        const matchedText = text.substring(
-            matchIndex,
-            matchIndex + normalizedQuery.length
-        );
-        const escapedMatch = this.escapeHtml(matchedText).replace(
-            /[.*+?^${}()|[\]\\]/g,
-            '\\$&'
-        );
-
-        // Create regex and highlight
-        const regex = new RegExp(`(${escapedMatch})`, 'gi');
-        return escapedText.replace(regex, '<strong>$1</strong>');
-    }
-
-    /**
-     * Render special states and options
-     */
     renderNoResults(query) {
-        // Use translated text for no results message
         return `<div class="search-no-results">
             <div class="search-no-results-icon">🔍</div>
             <div class="search-no-results-text">${trans('search_index.noResult')}</div>
@@ -352,141 +164,8 @@ export default class extends Controller {
         </div>`;
     }
 
-    /**
-     * Show loading and error states
-     */
-    showLoadingState() {
-        if (!this.hasResultsTarget) return;
-        this.resultsTarget.innerHTML = `<div class="search-loading">
-            <div class="search-loading-spinner"></div>
-            <div class="search-loading-text">${trans('app.search.searching')}</div>
-        </div>`;
-        this.showDropdown();
-    }
-
-    showErrorState(message) {
-        if (!this.hasResultsTarget) return;
-        this.resultsTarget.innerHTML = `<div class="search-error">
-            <div class="search-error-icon">⚠️</div>
-            <div class="search-error-text">${this.escapeHtml(message)}</div>
-        </div>`;
-        this.showDropdown();
-    }
-
-    /**
-     * Show the dropdown
-     */
-    showDropdown() {
-        if (!this.hasDropdownTarget) return;
-
-        this.isOpen = true;
-        this.dropdownTarget.classList.add('show');
-        this.element.classList.add('search-open');
-
-        if (this.hasInputTarget) {
-            this.inputTarget.setAttribute('aria-expanded', 'true');
-        }
-    }
-
-    /**
-     * Hide the dropdown
-     */
-    hideDropdown() {
-        if (!this.hasDropdownTarget) return;
-
-        this.isOpen = false;
-        this.dropdownTarget.classList.remove('show');
-        this.element.classList.remove('search-open');
-        this.selectedIndex = -1;
-
-        if (this.hasInputTarget) {
-            this.inputTarget.setAttribute('aria-expanded', 'false');
-        }
-
-        // Remove selection highlighting
-        this.clearSelection();
-    }
-
-    /**
-     * Handle keyboard navigation
-     */
-    handleKeydown(event) {
-        if (!this.isOpen) return;
-
-        const items = this.resultsTarget.querySelectorAll(
-            '.search-result-item, .search-show-more'
-        );
-
-        switch (event.key) {
-            case 'ArrowDown':
-                event.preventDefault();
-                this.selectedIndex = Math.min(
-                    this.selectedIndex + 1,
-                    items.length - 1
-                );
-                this.updateSelection(items);
-                break;
-
-            case 'ArrowUp':
-                event.preventDefault();
-                this.selectedIndex = Math.max(this.selectedIndex - 1, -1);
-                this.updateSelection(items);
-                break;
-
-            case 'Enter':
-                event.preventDefault();
-                if (this.selectedIndex >= 0 && items[this.selectedIndex]) {
-                    this.selectItem(items[this.selectedIndex]);
-                } else {
-                    // No selection, go to search results page
-                    this.showMoreResults();
-                }
-                break;
-
-            case 'Escape':
-                event.preventDefault();
-                this.hideDropdown();
-                break;
-        }
-    }
-
-    /**
-     * Update visual selection highlighting
-     */
-    updateSelection(items) {
-        // Clear previous selection
-        this.clearSelection();
-
-        // Highlight current selection
-        if (this.selectedIndex >= 0 && items[this.selectedIndex]) {
-            items[this.selectedIndex].classList.add('selected');
-
-            // Scroll into view if needed
-            items[this.selectedIndex].scrollIntoView({
-                block: 'nearest',
-                behavior: 'smooth',
-            });
-        }
-    }
-
-    /**
-     * Clear selection highlighting
-     */
-    clearSelection() {
-        const selectedItems = this.resultsTarget.querySelectorAll('.selected');
-        selectedItems.forEach((item) => item.classList.remove('selected'));
-    }
-
-    /**
-     * Set up click handlers for result items
-     */
-    setupResultClickHandlers() {
-        const resultItems = this.resultsTarget.querySelectorAll(
-            '.search-result-item'
-        );
-        resultItems.forEach((item) => {
-            item.addEventListener('click', () => this.selectItem(item));
-        });
+    handleEmptyEnter() {
+        this.showMoreResults();
     }
 
     /**
@@ -499,7 +178,6 @@ export default class extends Controller {
 
         if (!type || !slug) return;
 
-        // Determine the route based on type
         let routeName;
         let routeParams = { slug: slug };
 
@@ -514,13 +192,11 @@ export default class extends Controller {
                 break;
             case 'user':
                 routeName = 'user_show';
-                // User routes only need slug
                 break;
             default:
                 return;
         }
 
-        // Navigate to the selected item
         const url = this.generateRoute(routeName, routeParams);
         if (url) {
             window.location.href = url;
@@ -566,12 +242,6 @@ export default class extends Controller {
         return null;
     }
 
-    escapeHtml(text) {
-        const div = document.createElement('div');
-        div.textContent = text;
-        return div.innerHTML;
-    }
-
     /**
      * Clear the search input and hide dropdown
      */
@@ -597,6 +267,16 @@ export default class extends Controller {
                 'has-content',
                 this.inputTarget.value.trim().length > 0
             );
+        }
+    }
+
+    /**
+     * Handle input focus to clear placeholder behavior
+     */
+    handleFocus(event) {
+        const input = event.target;
+        if (input.value.trim().length >= this.minLengthValue) {
+            this.search(event);
         }
     }
